rename module and wrap decorator
