+++ deepspeed/__init__.py
@@ -37,7 +37,7 @@ from .module_inject import replace_transformer_layer, revert_transformer_layer
 from .utils import log_dist, OnDevice, logger
 from .comm.comm import init_distributed
 
-from .runtime import zero
+from .runtime import compiler, zero
 from .runtime import DeepSpeedOptimizer, ZeROOptimizer
 
 from .pipe import PipelineModule
